onboarding-langraph/requirements.txt
langchain
langgraph
langchain-community
tavily-python
langchain-core~=0.3.15
langchain_openai
langchain_core
langgraph
pydantic
python-dotenv
json5
copilotkit
langgraph-cli
